Update the addon template so it doesn't ship a plugin for documentation purposes.
Background:
At some point in like 2014, NVDA included support for documentation via an entry in the manifest. The version of the addon template this uses is quite old, and used a hack which shipped a plugin with NVDA, just to put a "help" entry in the help menu.
This pull request updates the addon template to not include this. It'll improve performance of NVDA loading, because only one plugin is loaded, not two. The readme needs updated, but that should be discussed as to what should be put in there.

Also, improvements were made to the way the plugin puts dependencies in the build chain. All .py and .wav files under globalPlugins are included now, because we might as well.